Rename BuildFolder to BuildFolderRule to keep consistent naming
