+++ core/src/test/groovy/com/novoda/gradle/release/test/BuildFolderRule.groovy
@@ -1,22 +1,22 @@
-package com.novoda.gradle.release
+package com.novoda.gradle.release.test
 
 
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
 
-class BuildFolder implements TestRule {
+class BuildFolderRule implements TestRule {
 
     private File rootDir
 
-    BuildFolder(String path = '') {
+    BuildFolderRule(String path = '') {
         File buildDir = new File(getResource('.').file).parentFile.parentFile.parentFile
         assert buildDir.path.endsWith('core/build')
         rootDir = new File(buildDir, path)
     }
 
     private static URL getResource(String resourceName) {
-        ClassLoader loader = Thread.currentThread().getContextClassLoader() ?: BuildFolder.class.getClassLoader()
+        ClassLoader loader = Thread.currentThread().getContextClassLoader() ?: BuildFolderRule.class.getClassLoader()
         URL url = loader.getResource(resourceName)
         if (url == null) {
             throw new IllegalArgumentException("resource ${resourceName} not found.")
